Refs #30899 -- Moved _lazy_re_compile() to the django.utils.regex_helper.
